requirements.txt
# Core dependencies
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.0.3
SQLAlchemy==2.0.49
PyYAML>=6.0
gradio
pydantic<2.0
spacy>=3.5.3
thinc>=8.2.0
httpx==0.23.0
httpcore==0.15.0

# File processing
PyPDF2==2.11.1
python-docx==0.8.11
markdown==3.4.1

# LLM integration
openai

# Vector database
chromadb==1.5.7

# Embedding models
sentence-transformers
numpy<2.0
huggingface-hub==0.21.0
torch>=2.2.0
transformers<5.0

# Utilities
python-dotenv==0.20.0
requests==2.28.1

# Development
pytest==7.1.2
black==22.3.0
flake8==4.0.1

# Performance monitoring
psutil==5.9.0

setuptools